requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
motor>=3.3.2
pymongo>=4.6.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.5.0
aiosmtplib>=3.0.1
python-dotenv>=1.0.0
jinja2>=3.1.2
email-validator>=2.1.0
bcrypt>=4.0.0
